refactor: fix linting
