+++ packages/button-primary/src/button-primary.js
@@ -6,16 +6,20 @@ import styles from "./styles";
 
 const ButtonPrimary = ({ onPress, style, title }) => {
   const cleanTitle = cleanUpTitle(title);
-	return (
-	  <View style={[styles.listErrorButtonContainer, style]}>
-	    <TouchableOpacity accessible accessibilityLabel={cleanTitle} onPress={onPress}>
-	      <View style={styles.listErrorButton}>
-	        <Text style={styles.listErrorButtonText}>{cleanTitle}</Text>
-	      </View>
-	    </TouchableOpacity>
-	  </View>
-	);
-}
+  return (
+    <View style={[styles.listErrorButtonContainer, style]}>
+      <TouchableOpacity
+        accessible
+        accessibilityLabel={cleanTitle}
+        onPress={onPress}
+      >
+        <View style={styles.listErrorButton}>
+          <Text style={styles.listErrorButtonText}>{cleanTitle}</Text>
+        </View>
+      </TouchableOpacity>
+    </View>
+  );
+};
 
 ButtonPrimary.propTypes = propTypes;
 ButtonPrimary.defaultProps = defaultProps;
