+++ packages/button-primary/src/styles/index.android.js
@@ -1,17 +1,34 @@
 import { StyleSheet } from "react-native";
-import sharedStyles from "./shared";
+import { colours, fonts, fontSizes } from "@times-components/styleguide";
 
 const styles = StyleSheet.create({
-  ...sharedStyles,
-  listErrorButton: {
-  	...sharedStyles.listErrorButton,
-  	borderRadius: 4,
-  	height: 36,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.2,
-    elevation: 10,
-    position:'relative'
+  button: {
+    alignItems: "center",
+    backgroundColor: "#FFFFFF",
+    flex: 1,
+    height: 36,
+    justifyContent: "center",
+    width: "100%"
+  },
+  buttonContainer: {
+    alignItems: "center",
+    backgroundColor: colours.functional.action,
+    borderRadius: 4,
+    elevation: 3,
+    height: 36,
+    justifyContent: "center",
+    shadowColor: "#000000",
+    shadowOffset: {
+      width: 0,
+      height: 3
+    },
+    shadowOpacity: 1.0,
+    shadowRadius: 5
+  },
+  textStyle: {
+    color: "#FFFFFF",
+    fontFamily: fonts.supporting,
+    fontSize: fontSizes.buttonPrimary
   }
 });
